+++ src/main/java/org/scion/jpan/ScionService.java
@@ -30,6 +30,7 @@ import static org.scion.jpan.Constants.PROPERTY_DAEMON;
 import static org.scion.jpan.Constants.PROPERTY_DNS_SEARCH_DOMAINS;
 import static org.scion.jpan.Constants.PROPERTY_USE_OS_SEARCH_DOMAINS;
 
+import com.google.protobuf.Empty;
 import io.grpc.*;
 import java.io.IOException;
 import java.net.InetAddress;
@@ -39,7 +40,6 @@ import java.nio.file.Paths;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.stream.Collectors;
 import org.scion.jpan.internal.*;
 import org.scion.jpan.proto.control_plane.SegmentLookupServiceGrpc;
 import org.scion.jpan.proto.daemon.Daemon;
@@ -75,6 +75,7 @@ public class ScionService {
   private final ScionBootstrapper bootstrapper;
   private final DaemonServiceGrpc.DaemonServiceBlockingStub daemonStub;
   private final SegmentLookupServiceGrpc.SegmentLookupServiceBlockingStub segmentStub;
+  private LocalTopology.DispatcherPortRange portRange;
 
   private final boolean minimizeRequests;
   private final ManagedChannel channel;
@@ -645,14 +646,28 @@ public class ScionService {
     }
   }
 
-  List<String> getBorderRouterStrings() {
-    if (daemonStub != null) {
-      return getInterfaces().values().stream()
-          .map(i -> i.getAddress().getAddress())
-          .collect(Collectors.toList());
-    } else {
-      return bootstrapper.getLocalTopology().getBorderRouterAddresses();
+  LocalTopology.DispatcherPortRange getLocalPortRange() {
+    if (portRange == null) {
+      if (bootstrapper != null) {
+        portRange = bootstrapper.getLocalTopology().getPortRange();
+      } else if (daemonStub != null) {
+        // try daemon
+        Daemon.PortRangeResponse response;
+        try {
+          response = daemonStub.portRange(Empty.getDefaultInstance());
+          portRange =
+              LocalTopology.DispatcherPortRange.create(
+                  response.getDispatchedPortStart(), response.getDispatchedPortEnd());
+        } catch (StatusRuntimeException e) {
+          LOG.warn("ERROR getting port range from daemon: {}", e.getMessage());
+          // Daemon doesn't support port range.
+          portRange = LocalTopology.DispatcherPortRange.createEmpty();
+        }
+      } else {
+        portRange = LocalTopology.DispatcherPortRange.createAll();
+      }
     }
+    return portRange;
   }
 
   InetSocketAddress getBorderRouterAddress(int interfaceID) {