+++ src/main/java/org/scion/jpan/internal/LocalTopology.java
@@ -17,10 +17,12 @@ package org.scion.jpan.internal;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import org.scion.jpan.Constants;
 import org.scion.jpan.ScionRuntimeException;
 import org.scion.jpan.ScionUtil;
 
@@ -33,6 +35,7 @@ public class LocalTopology {
   private String localIsdAs;
   private boolean isCoreAs;
   private int localMtu;
+  private DispatcherPortRange portRange;
 
   public static synchronized LocalTopology create(String topologyFile) {
     LocalTopology topo = new LocalTopology();
@@ -69,14 +72,6 @@ public class LocalTopology {
       }
     }
     throw new ScionRuntimeException("No router found with interface ID " + interfaceId);
-  }
-
-  public List<String> getBorderRouterAddresses() {
-    List<String> result = new ArrayList<>();
-    for (BorderRouter br : borderRouters) {
-      result.add(br.internalAddress);
-    }
-    return result;
   }
 
   public int getMtu() {
@@ -123,6 +118,12 @@ public class LocalTopology {
           discoveryServices.add(new ServiceNode(e.getKey(), ds.get("addr").getAsString()));
         }
       }
+      JsonElement dispatchedPorts = o.get("dispatched_ports");
+      if (dispatchedPorts == null) {
+        portRange = DispatcherPortRange.createEmpty();
+      } else {
+        portRange = parsePortRange(dispatchedPorts.getAsString());
+      }
       JsonArray attr = safeGet(o, "attributes").getAsJsonArray();
       for (int i = 0; i < attr.size(); i++) {
         if ("core".equals(attr.get(i).getAsString())) {
@@ -130,6 +131,29 @@ public class LocalTopology {
         }
       }
     }
+  }
+
+  private static DispatcherPortRange parsePortRange(String v) {
+    if ("-".equals(v)) {
+      return DispatcherPortRange.createEmpty();
+    } else if ("all".equalsIgnoreCase(v)) {
+      return DispatcherPortRange.createAll();
+    } else {
+      String[] sa = v.split("-");
+      if (sa.length != 2) {
+        throw new ScionRuntimeException("Illegal expression in topo file dispatched_ports: " + v);
+      }
+      int portMin = Integer.parseInt(sa[0]);
+      int portMax = Integer.parseInt(sa[1]);
+      if (portMin < 1 || portMax < 1 || portMax > 65535 || portMin > portMax) {
+        throw new ScionRuntimeException("Illegal port values in topo file dispatched_ports: " + v);
+      }
+      return DispatcherPortRange.create(portMin, portMax);
+    }
+  }
+
+  public DispatcherPortRange getPortRange() {
+    return portRange;
   }
 
   @Override
@@ -229,6 +253,48 @@ public class LocalTopology {
     @Override
     public String toString() {
       return "{" + "name='" + name + '\'' + ", ipString='" + ipString + '\'' + '}';
+    }
+  }
+
+  public static class DispatcherPortRange {
+    private final int portMin;
+    private final int portMax;
+
+    private DispatcherPortRange(int min, int max) {
+      portMin = min;
+      portMax = max;
+    }
+
+    public static DispatcherPortRange create(int min, int max) {
+      return new DispatcherPortRange(min, max);
+    }
+
+    public static DispatcherPortRange createAll() {
+      return new DispatcherPortRange(1, 65535);
+    }
+
+    public static DispatcherPortRange createEmpty() {
+      return new DispatcherPortRange(-1, -2);
+    }
+
+    public boolean hasPortRange() {
+      return portMin >= 1 && portMax <= 65535 && portMax >= portMin;
+    }
+
+    public InetSocketAddress mapToLocalPort(InetSocketAddress address) {
+      if (address.getPort() == Constants.SCMP_PORT
+          || (address.getPort() >= portMin && address.getPort() <= portMax)) {
+        return address;
+      }
+      return new InetSocketAddress(address.getAddress(), Constants.DISPATCHER_PORT);
+    }
+
+    public int getPortMin() {
+      return portMin;
+    }
+
+    public int getPortMax() {
+      return portMax;
     }
   }
 }