+++ src/test/java/org/scion/jpan/api/DatagramChannelApiServerTest.java
@@ -112,6 +112,7 @@ class DatagramChannelApiServerTest {
     // check that send(Path) does not internally require a ScionService.
     ScionService.closeDefault();
     try (ScionDatagramChannel channel = ScionDatagramChannel.open(null)) {
+      channel.bind(new InetSocketAddress("127.0.0.1", 12345));
       assertNull(channel.getService());
       ResponsePath path =
           PackageVisibilityHelper.createDummyResponsePath(
@@ -139,6 +140,7 @@ class DatagramChannelApiServerTest {
           return addr;
         });
     try (ScionDatagramChannel channel = ScionDatagramChannel.open(null, mock)) {
+      channel.bind(new InetSocketAddress("127.0.0.1", 12345));
       assertNull(channel.getService());
       ByteBuffer buffer = ByteBuffer.allocate(100);
       ScionSocketAddress responseAddress = channel.receive(buffer);