+++ src/test/java/org/scion/jpan/api/ScmpResponderTest.java
@@ -21,11 +21,14 @@ import java.net.*;
 import java.nio.ByteBuffer;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
 import java.util.function.Predicate;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.scion.jpan.*;
+import org.scion.jpan.internal.Shim;
 import org.scion.jpan.testutil.ManagedThread;
 import org.scion.jpan.testutil.ManagedThreadNews;
 import org.scion.jpan.testutil.MockNetwork;
@@ -34,10 +37,17 @@ import org.scion.jpan.testutil.MockScmpHandler;
 public class ScmpResponderTest {
   private static final ConcurrentLinkedQueue<String> errors = new ConcurrentLinkedQueue<>();
 
+  @BeforeAll
+  public static void beforeAll() {
+    System.setProperty(Constants.PROPERTY_SHIM, "false");
+    Shim.uninstall();
+  }
+
   @AfterAll
   public static void afterAll() {
     // Defensive clean up
     ScionService.closeDefault();
+    System.clearProperty(Constants.PROPERTY_SHIM);
   }
 
   @AfterEach